Create 3. Design
This script is different to the previous two in that it doesn't use a trial value of Uo. Instead the design decisions are used to calculate both the area and overall heat transfer coefficient. The nozzle pressure drops have been included for the first time - they didn't affect the results.